+++ layout/graphyz_test.go
@@ -1,4 +1,4 @@
-package generator
+package layout
 
 import (
 	"testing"
